fixed UIs (#14)
* anvu fixed set_alarm_page UI, fixed icons on home_page

* tanphat fixed 2 bugs: navbar bug and back_button in chart_page,dart

---------

+++ lib/widgets/chart.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 class SleepChart extends StatelessWidget {
   final List<String> daysOfWeek;
-  final List<int> sleepHours;
+  final List<double> sleepHours;
   final EdgeInsets padding;
 
   SleepChart({
